Modificacoes estruturais e no codigo dos handlers das excecoes
